API Reference

Gomen API

Games

Gomen.Game — Type.

Game(s, t)

A two-player, two-strategy, symmetric game. The games are characterized by the diagonal elements of a payoff matrix

$\begin{bmatrix} 1 & S \\ T & 0 \end{bmatrix}$

with $-0.5 \leq S \leq 0.5$ and $0.5 \leq T \leq 1.5$,

source

Gomen.play — Function.

play(g::Game, a, b)

Play the game with the strategies a and b, returning the payoff recieved by agent playing a.

source