API Reference

Gomen API

Games

Gomen.Game — Type.

Game(s, t)

A two-player, two-strategy, symmetric game. The games are characterized by the diagonal elements of a payoff matrix

\[\begin{bmatrix} 1 & S \\ T & 0 \end{bmatrix}\]

with $-0.5 \\leq S \\leq 0.5$ and $0.5 \\leq T \\leq 1.5$,

source

Gomen.play — Function.

play(g::Game, a, b)

Play the game with the strategies a and b, returning the payoff recieved by agent playing a.

source

Rules

Gomen.AbstractRule — Type.

AbstractRule

A supertype for all decision rules.

source

Gomen.apply — Function.

apply(r::AbstractRule, dp)

Determine the probability an agent will switch its strategy based on the rule and the difference in payoffs of the two strategies.

source

Gomen.Sigmoid — Type.

Sigmoid(β)

A rule specifying that the agent will switch its strategy with probability

\[P(dp) = \frac{1}{1 + e^{-\beta dp}}\]

where $dp$ is the difference in the two payoffs.

source

Gomen.Heaviside — Type.

Heaviside(ϵ)

A rule specifying that the agent will switch its strategy with probability

\[P(dp) = \begin{cases} 0.0, & dp \leq -\epsilon \\ 0.5, & -\epsilon < dp < \epsilon \\ 1.0, & \epsilon < dp \end{cases}\]

where $dp$ is the difference in the two payoffs.

source

Schemes

Gomen.AbstractScheme — Type.

AbstractScheme

A supertype for all schemes, mechanisms for agents to choose their next strategy.

source

Gomen.decide — Function.

decide(scheme, ss::AbstractVector{Int}, ps::AbstractArray{Float64,2})

Use the scheme to decide what each agent's strategy will be in the next timestep. Here ss is a vector of the agents' current strategies (one for each agent) and ps is a matrix of payoffs. Each column of ps should correspond to an agent, and each row is a corresponding strategy.

source

decide(scheme, ss::AbstractVector{Int}, dps::AbstractVector{Float64})

Use the scheme to decide what each agent's strategy will be in the next time step. Here ss is a vector of the agents' current strategies (one for each agent) and dps is a vector of the difference between the payoff for the strategy the agent didn't play and the payoff they actually received.

decide(scheme, s::Int, dp::Float64)

Use the scheme to decide what an agent's strategy will be in the next time step. Here s is the agent's current strategy and dp is the difference between the payoff for the alternative strategy and the actual payoff they received.

source

Gomen.CounterFactual — Type.

CounterFactural([rule = Sigmoid()])

The "counter-factual" update scheme. Under this scheme, the agent will switch their strategy according to some probability. The probability the agent will switch is determined by applying the scheme's rule to the payoff difference.

source