API Reference

Gomen API

Games

Gomen.Game — Type.

Game(s, t)

A two-player, two-strategy, symmetric game. The games are characterized by the diagonal elements of a payoff matrix

\[\begin{bmatrix} 1 & S \\ T & 0 \end{bmatrix}\]

with $-0.5 \leq S \leq 0.5$ and $0.5 \leq T \leq 1.5$,

source

Gomen.play — Function.

play(g::Game, a, b)

Play the game with the strategies a and b, returning the payoff recieved by agent playing a.

source

play(arena, strategies)

Play one round of the game in the arena, returning the agents' next strategies.

source

play(arena, rounds)

Play the game in the arena repeatedly with the agents starting out with a random initial strategy. The result is a matrix of the agents strategies with rounds many rows, one for each round played.

source

play(arena, rounds, replicates)

Starting from replicates many random initial strategies, play the game in the arena rounds many times. The result is a 3-D array of the agents strategies with the first two dimensions of size replicates and rounds, respectively.

source

Gomen.Games — Type.

Games(sstep, tstep)

An iterator over the games with $S$ and $T$ parameters varied over their ranges with the provided step sizes.

source

Rules

Gomen.AbstractRule — Type.

AbstractRule

A supertype for all decision rules.

source

Gomen.apply — Function.

apply(r::AbstractRule, dp)

Determine the probability an agent will switch its strategy based on the rule and the difference in payoffs of the two strategies.

source

Gomen.Sigmoid — Type.

Sigmoid(β)

A rule specifying that the agent will switch its strategy with probability

\[P(dp) = \frac{1}{1 + e^{-\beta dp}}\]

where $dp$ is the difference in the two payoffs.

source

Gomen.Heaviside — Type.

Heaviside(ϵ)

A rule specifying that the agent will switch its strategy with probability

\[P(dp) = \begin{cases} 0.0, & dp \leq -\epsilon \\ 0.5, & -\epsilon < dp < \epsilon \\ 1.0, & \epsilon < dp \end{cases}\]

where $dp$ is the difference in the two payoffs.

source

Schemes

Gomen.AbstractScheme — Type.

AbstractScheme

A supertype for all schemes, mechanisms for agents to choose their next strategy.

source

Gomen.decide — Function.

decide(scheme, ss::AbstractVector{Int}, ps::AbstractArray{Float64,2})

Use the scheme to decide what each agent's strategy will be in the next timestep. Here ss is a vector of the agents' current strategies (one for each agent) and ps is a matrix of payoffs. Each column of ps should correspond to an agent, and each row is a corresponding strategy.

source

decide(scheme, ss::AbstractVector{Int}, dps::AbstractVector{Float64})

Use the scheme to decide what each agent's strategy will be in the next time step. Here ss is a vector of the agents' current strategies (one for each agent) and dps is a vector of the difference between the payoff for the strategy the agent didn't play and the payoff they actually received.

decide(scheme, s::Int, dp::Float64)

Use the scheme to decide what an agent's strategy will be in the next time step. Here s is the agent's current strategy and dp is the difference between the payoff for the alternative strategy and the actual payoff they received.

source

Gomen.CounterFactual — Type.

CounterFactural([rule = Sigmoid()])

The "counter-factual" update scheme. Under this scheme, the agent will switch their strategy according to some probability. The probability the agent will switch is determined by applying the scheme's rule to the payoff difference.

source

Arena

Gomen.AbstractArena — Type.

AbstractArena{Graph, Scheme}

A supertype for all game-playing arenas.

source

Gomen.game — Function.

game(arena)

The game the agents in the arena play

source

Gomen.graph — Function.

graph(arena)

The graph on which the agents play their games

source

Gomen.scheme — Function.

scheme(arena)

The scheme the agents use to determine their next strategy

source

Base.length — Function.

length(arena)

The number of agents in the arena

source

LightGraphs.edges — Function.

edges(arena)

The edges connecting agents in the arena

source

LightGraphs.neighbors — Function.

neighbors(arena, agent)

A vector of the agent's neighbors

source

Gomen.Arena — Type.

Arena(game, graph, scheme) <: AbstractArena

An arena in which the agents, situated on the vertices of a graph, play a game with their neighbors. Every pair of agents plays the same game. With each round of play, the agents use the scheme to decide what their next strategy will be.

In this arena, the agents accumulate payoff by playing with every neighbor. The decision of what strategy to use in the next time step is based on the difference between the total payoffs for each of the agents' strategies.

source

Gomen.payoffs — Function.

payoffs(arena, strategies)

Determine the payoff of each agents' possible strategies given the strategies of all of the other agents.

source